[feat]: von-4 - refacto rofim customisation of the app
