+++ frontend/index.html
@@ -12,6 +12,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/rofim/main.tsx"></script>
+    <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
