+++ frontend/src/App.spec.tsx
@@ -4,7 +4,7 @@ import { PropsWithChildren } from 'react';
 import App from './App';
 
 // Mock the page components to make text assertions easy
-vi.mock('./pages/LandingPage', () => ({ default: () => <div>Landing Page</div> }));
+vi.mock('./pages/WaitingRoom', () => ({ default: () => <div>Waiting Room Page</div> }));
 vi.mock('./pages/GoodBye/index', () => ({ default: () => <div>GoodBye Page</div> }));
 vi.mock('./pages/UnsupportedBrowserPage', () => ({
   default: () => <div>Unsupported Browser</div>,
@@ -14,13 +14,11 @@ vi.mock('./components/RedirectToWaitingRoom', () => ({
   default: ({ children }: PropsWithChildren) => children,
 }));
 
-describe('App routing', () => {
-  it('renders LandingPage on unknown route', () => {
-    window.history.pushState({}, '', '/unknown');
-    render(<App />);
-    expect(screen.getByText(/Landing Page/i)).toBeInTheDocument();
-  });
+vi.mock('@rofim/context/RofimContext', () => ({
+  default: ({ children }: PropsWithChildren) => children,
+}));
 
+describe('App routing', () => {
   it('renders GoodBye page on /goodbye', () => {
     window.history.pushState({}, '', '/goodbye');
     render(<App />);
