+++ frontend/src/components/DeviceAccessAlert/DeviceAccessAlert.tsx
@@ -1,10 +1,8 @@
 import { ReactElement } from 'react';
 import { useTranslation } from 'react-i18next';
 import { DEVICE_ACCESS_STATUS } from '../../utils/constants';
-import { isWebKit } from '../../utils/util';
 import Alert from '@mui/material/Alert';
 import Dialog from '@mui/material/Dialog';
-import Box from '@mui/material/Box';
 import Stack from '@mui/material/Stack';
 import AlertTitle from '@mui/material/AlertTitle';
 
@@ -25,52 +23,32 @@ const DeviceAccessAlert = ({ accessStatus }: DeviceAccessAlertProps): ReactEleme
     accessStatus === DEVICE_ACCESS_STATUS.PENDING
       ? t('deviceAccessAlert.askDeviceMessage')
       : t('deviceAccessAlert.deniedDeviceMessage');
-  const imgToDisplay =
-    accessStatus === DEVICE_ACCESS_STATUS.PENDING
-      ? '/images/access-dialog-pending.png'
-      : '/images/access-dialog-rejected.png';
+
   const severity = accessStatus === DEVICE_ACCESS_STATUS.PENDING ? 'success' : 'error';
 
   return (
-    !isWebKit() && (
-      <Stack
-        sx={{
-          width: '40%',
-          position: 'absolute',
-          left: '50%',
-          top: '50%',
-          transform: 'translate(-50%, -50%)',
-          zIndex: 10,
-          alignItems: 'center',
-          padding: 2,
-          borderRadius: 2,
-        }}
-        spacing={3}
-      >
-        {accessStatus && (
-          <Dialog open>
-            <Alert severity={severity} variant="outlined" sx={{ width: '100%' }}>
-              <AlertTitle>{messageToDisplay}</AlertTitle>
-              <Box
-                sx={{
-                  display: 'flex',
-                  justifyContent: 'center',
-                  alignItems: 'center',
-                  width: '100%',
-                }}
-              >
-                <Box
-                  component="img"
-                  src={imgToDisplay}
-                  alt={t('deviceAccessAlert.imageAlt')}
-                  sx={{ maxWidth: 300, borderRadius: 1 }}
-                />
-              </Box>
-            </Alert>
-          </Dialog>
-        )}
-      </Stack>
-    )
+    <Stack
+      sx={{
+        width: '40%',
+        position: 'absolute',
+        left: '50%',
+        top: '50%',
+        transform: 'translate(-50%, -50%)',
+        zIndex: 10,
+        alignItems: 'center',
+        padding: 2,
+        borderRadius: 2,
+      }}
+      spacing={3}
+    >
+      {accessStatus && (
+        <Dialog open>
+          <Alert severity={severity} sx={{ width: '100%' }}>
+            <AlertTitle>{messageToDisplay}</AlertTitle>
+          </Alert>
+        </Dialog>
+      )}
+    </Stack>
   );
 };
