+++ frontend/src/css/App.css
@@ -1,3 +1,7 @@
+* {
+  font-family: 'Inter', sans-serif;
+}
+
 #root {
   margin: 0 auto;
   text-align: center;
