+++ frontend/src/pages/WaitingRoom/WaitingRoom.tsx
@@ -1,18 +1,16 @@
 import { FC } from 'react';
 import Box from '@mui/material/Box';
 import PageLayout from '@ui/PageLayout';
-import Banner from '@components/Banner';
-import Footer from '@components/Footer/Footer';
 import ControlPanel from '@components/WaitingRoom/ControlPanel';
 import VideoContainer from '@components/WaitingRoom/VideoContainer';
-import UsernameInput from '@components/WaitingRoom/UserNameInput';
 import DeviceAccessAlert from '@components/DeviceAccessAlert';
 import { DEVICE_ACCESS_STATUS } from '@utils/constants';
 import backgroundEffectsDialog$ from '@Context/BackgroundEffectsDialog';
 import precallNetworkTestDialog$ from '@Context/PrecallNetworkTestDialog';
 import VideoContainerSkeleton from '@components/WaitingRoom/VideoContainer/VideoContainer.skeleton';
-import UsernameInputSkeleton from '@components/WaitingRoom/UserNameInput/UserNameInput.skeleton';
 import useWaitingRoom from '@hooks/useWaitingRoom';
+import useEquipmentsTestRoom from '@rofim/hooks/useEquipmentsTestRoom';
+import JoinRoomButton from '@rofim/components/JoinRoomButton/JoinRoomButton';
 
 /**
  * WaitingRoom Component
@@ -36,7 +34,6 @@ const WaitingRoom: FC = () => {
     openVideoInput,
     openAudioOutput,
     username,
-    setUsername,
     accessStatus,
     isRoomReady,
     handleAudioInputOpen,
@@ -45,15 +42,13 @@ const WaitingRoom: FC = () => {
     handleClose,
   } = useWaitingRoom();
 
+  const { isLoading, patientId, waitingRoom, handleJoinClick } = useEquipmentsTestRoom();
+
   return (
     <backgroundEffectsDialog$.Provider>
       <precallNetworkTestDialog$.Provider>
         <Box data-testid="waitingRoom">
           <PageLayout>
-            <PageLayout.Banner>
-              <Banner />
-            </PageLayout.Banner>
-
             <PageLayout.Left>
               <Box
                 className={`relative flex flex-col sm:inline-flex h-auto max-w-full sm:h-100 animate-fade-in`}
@@ -61,7 +56,6 @@ const WaitingRoom: FC = () => {
                 {isRoomReady && (
                   <>
                     <VideoContainer username={username} />
-
                     <ControlPanel
                       handleAudioInputOpen={handleAudioInputOpen}
                       handleVideoInputOpen={handleVideoInputOpen}
@@ -72,28 +66,19 @@ const WaitingRoom: FC = () => {
                       openAudioOutput={openAudioOutput}
                       anchorEl={anchorEl}
                     />
+                    <JoinRoomButton
+                      isLoading={isLoading}
+                      isDisabled={!username && isLoading}
+                      patientId={patientId}
+                      waitingRoom={waitingRoom}
+                      onClick={handleJoinClick}
+                    />
                   </>
                 )}
 
                 {!isRoomReady && <VideoContainerSkeleton />}
               </Box>
             </PageLayout.Left>
-
-            <PageLayout.Right>
-              {isRoomReady && (
-                <UsernameInput
-                  className={`flex-col sm:inline-flex h-auto sm:h-100 animate-fade-in`}
-                  username={username}
-                  setUsername={setUsername}
-                />
-              )}
-
-              {!isRoomReady && <UsernameInputSkeleton />}
-            </PageLayout.Right>
-
-            <PageLayout.Footer>
-              <Footer />
-            </PageLayout.Footer>
           </PageLayout>
           {accessStatus !== DEVICE_ACCESS_STATUS.ACCEPTED && (
             <DeviceAccessAlert accessStatus={accessStatus} />
