--- frontend/src/rofim/components/DeviceAccessAlert/index.tsx
@@ -1,3 +1,0 @@
-import DeviceAccessAlert from './DeviceAccessAlert';
-
-export default DeviceAccessAlert;
