--- frontend/src/rofim/css/global.css
@@ -1,3 +1,0 @@
-* {
-  font-family: 'Inter', sans-serif;
-}
