+++ frontend/src/rofim/environments/environment.local.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @cspell/spellchecker */
 import { IEnvironment } from './environment.interface';
 
 const environment: IEnvironment = {
@@ -6,10 +5,6 @@ const environment: IEnvironment = {
   production: false,
   apiUrl: 'http://localhost:3000',
   wsUrl: 'ws://localhost:3000',
-  i18n: {
-    defaultLanguage: 'fr',
-    supportedLanguages: ['de', 'en', 'es', 'fr', 'it', 'pt'],
-  },
   matomo: {
     url: '//matomo.staging.rofimoncloud.com/',
     siteId: '1',
