+++ frontend/src/rofim/environments/environment.preprod.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @cspell/spellchecker */
 import { IEnvironment } from './environment.interface';
 
 const environment: IEnvironment = {
@@ -6,10 +5,6 @@ const environment: IEnvironment = {
   production: true,
   apiUrl: 'https://api.preprod.rofim.doctor',
   wsUrl: 'wss://api.preprod.rofim.doctor',
-  i18n: {
-    defaultLanguage: 'fr',
-    supportedLanguages: ['de', 'en', 'es', 'fr', 'it', 'pt'],
-  },
   matomo: {
     url: '//matomo.preprod.rofim.doctor/',
     siteId: '1',
