+++ frontend/src/rofim/hooks/useRofimMeeting.ts
@@ -1,15 +1,11 @@
 /* eslint-disable @cspell/spellchecker */
-/* eslint @typescript-eslint/no-floating-promises: 0 */
-
 import { useEffect } from 'react';
-import VonageRoom from '../../../pages/MeetingRoom';
-import RofimApiService, { WaitingRoomStatus } from '../../api/rofimApi';
-import { getRofimSession } from '../../utils/session';
-import './room.css';
-import useWebSocket from '../../hooks/useWebSocket';
-import useSessionContext from '../../../hooks/useSessionContext';
+import RofimApiService, { WaitingRoomStatus } from '../api/rofimApi';
+import { getRofimSession } from '../utils/session';
+import useWebSocket from './useWebSocket';
+import useSessionContext from '../../hooks/useSessionContext';
 
-const Room = () => {
+const useRofimMeeting = () => {
   const rofimSession = getRofimSession();
   const patientId = rofimSession?.patientId;
   const type = rofimSession?.type;
@@ -22,17 +18,17 @@ const Room = () => {
         // TODO: a refacto quand on aura plus vonageV1
         // Pour laisser le temps au WS de se reconnecter avant d'appeler l'API
         const timeout = setTimeout(() => {
-          RofimApiService.updateTeleconsultationStatus(WaitingRoomStatus.Progress);
+          void RofimApiService.updateTeleconsultationStatus(WaitingRoomStatus.Progress);
         }, 5000);
         return () => clearTimeout(timeout);
       }
 
       if (subscriberWrappers.length === 0) {
         // this is for the doctor
-        RofimApiService.doctorJoinVisio();
+        void RofimApiService.doctorJoinVisio();
         // periodically try to join the visio in case of disconnection
         const interval = setInterval(() => {
-          RofimApiService.doctorJoinVisio();
+          void RofimApiService.doctorJoinVisio();
         }, 15000);
 
         return () => clearInterval(interval);
@@ -40,8 +36,6 @@ const Room = () => {
     }
     return () => {};
   }, [patientId, type, isSocketConnected, subscriberWrappers.length]);
-
-  return <VonageRoom />;
 };
 
-export default Room;
+export default useRofimMeeting;
