--- frontend/src/rofim/pages/ErrorPage.tsx
@@ -1,5 +1,0 @@
-const ErrorPage = () => {
-  return <div className="size-full bg-white">Error Page TODO</div>;
-};
-
-export default ErrorPage;
